chore: cleaner architecture, added sidebar.jsx to unweight layout.jsx
