+++ metricshub-web/react/src/build-info.json
@@ -1,3 +1,3 @@
 {
-  "buildTime": "2025-09-12T14:11:23.786Z"
+  "buildTime": "2025-09-15T08:43:36.403Z"
 }
