+++ metricshub-web/react/src/components/error-boundary.jsx
