+++ metricshub-web/react/src/components/status/status-details-menu.jsx
@@ -12,7 +12,7 @@ import {
 	Divider,
 } from "@mui/material";
 import InfoOutlinedIcon from "@mui/icons-material/InfoOutlined";
-import { useAppSelector } from "../hooks/store";
+import { useAppSelector } from "../../hooks/store";
 
 // key prettifier
 function prettifyKey(key = "") {
@@ -39,10 +39,10 @@ export default function StatusDetailsMenu() {
 	const rest =
 		data && typeof data === "object"
 			? Object.fromEntries(
-					Object.entries(data).filter(
-						([k]) => k !== "agentInfo" && k !== "status" && k !== "otelCollectorRunning",
-					),
-				)
+				Object.entries(data).filter(
+					([k]) => k !== "agentInfo" && k !== "status" && k !== "otelCollectorRunning",
+				),
+			)
 			: null;
 
 	return (
